Renaming Step files to include ".steps" (#5489)
* move the create wallet action into common step

* append "steps" to step files

* add proper tag CreatingWalletAccount feature file

* undo podfile.lock changes

* remove redundant imports

* missed renaming security checks step file.

* remove excessive timeout.

* removing redudant step
