+++ wdio/features/CreatingWalletAccount.feature
@@ -1,5 +1,6 @@
+@androidApp @ChainScenarios
+
 Feature: Creating account in wallet
-    @ChainScenarios
     Scenario: Import account
         Given I have imported my wallet
         # Given I have created my wallet
