+++ wdio/features/step-definitions/delete-wallet-modal-view.steps.js
